+++ wcfsetup/install/files/lib/system/package/plugin/LanguagePackageInstallationPlugin.class.php
@@ -254,6 +254,12 @@ class LanguagePackageInstallationPlugin extends AbstractXMLPackageInstallationPl
 	protected function findExistingItem(array $data) { }
 	
 	/**
+	 * @see	\wcf\system\package\plugin\IPackageInstallationPlugin::getDefaultFilename()
+	 */
+	public static function getDefaultFilename() {
+		return 'language/*.xml';
+	}
+	/**
 	 * @see	\wcf\system\package\plugin\IPackageInstallationPlugin::isValid()
 	 */
 	public static function isValid(PackageArchive $archive, $instruction) {