+++ wcfsetup/install/files/lib/system/package/plugin/SQLPackageInstallationPlugin.class.php
@@ -153,9 +153,20 @@ class SQLPackageInstallationPlugin extends AbstractPackageInstallationPlugin {
 	}
 	
 	/**
+	 * @see	\wcf\system\package\plugin\IPackageInstallationPlugin::getDefaultFilename()
+	 */
+	public static function getDefaultFilename() {
+		return 'install.sql';
+	}
+	
+	/**
 	 * @see	\wcf\system\package\plugin\IPackageInstallationPlugin::isValid()
 	 */
 	public static function isValid(PackageArchive $archive, $instruction) {
+		if (!$instruction) {
+			$defaultFilename = static::getDefaultFilename();
+		}
+		
 		if (preg_match('~\.sql$~', $instruction)) {
 			// check if file actually exists
 			try {